% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sanity.R
\name{.check_parameters}
\alias{.check_parameters}
\title{Check the hyperparameters}
\usage{
.check_parameters(lambdas, nonzeros)
}
\arguments{
\item{lambdas}{x}

\item{nonzeros}{x}
}
\value{
logical; TRUE or FALSE
}
\description{
Check the hyperparameters
}
