% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sanity.R
\name{.check_rda_class}
\alias{.check_rda_class}
\title{Check if a variable is a sRDA object}
\usage{
.check_rda_class(object)
}
\arguments{
\item{object}{any}
}
\value{
logical; TRUE or FALSE
}
\description{
Check if a variable is a sRDA object
}
