% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_penalty_parameters.R
\name{.get_cv_penalty_parameters}
\alias{.get_cv_penalty_parameters}
\title{Title}
\usage{
.get_cv_penalty_parameters(
  explanatory,
  response,
  penalization,
  lambdas,
  nonzeros,
  cv_n_folds,
  parallel,
  max_iteration,
  tolerance,
  seed
)
}
\arguments{
\item{explanatory}{x}

\item{response}{x}

\item{penalization}{x}

\item{lambdas}{x}

\item{nonzeros}{x}

\item{cv_n_folds}{x}

\item{parallel}{x}

\item{max_iteration}{x}

\item{tolerance}{x}

\item{seed}{x}
}
\value{
x
}
\description{
Title
}
\examples{
x
}
