% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cv_penalty_parameters.R
\name{.get_resampled_splits}
\alias{.get_resampled_splits}
\title{Title}
\usage{
.get_resampled_splits(X, Y, n_folds, seed)
}
\arguments{
\item{X}{x}

\item{Y}{x}

\item{n_folds}{x}

\item{seed}{x}
}
\value{
x
}
\description{
This is an internal function
}
\examples{
x
}
