% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ust.R
\name{.get_ust}
\alias{.get_ust}
\title{Get penalized weights with univariate soft-thresholding}
\usage{
.get_ust(x, y, nonzero)
}
\arguments{
\item{x}{
matrix of predictors
}

\item{y}{
response 
}

\item{nonzero}{The number of non-zero alpha weights of the explanatory latent variable (Xi)}
}
\value{
A matrix of penalized weights
}
\description{
This is an internal function. (UST; ridge = Inf)
}
\examples{
testing
}
