% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_paraellel_cv.R
\name{.run_parallel_cv}
\alias{.run_parallel_cv}
\title{Title}
\usage{
.run_parallel_cv(
  X,
  Y,
  lambdas,
  nonzeros,
  label,
  cv_n_folds,
  penalization,
  max_iteration,
  tolerance
)
}
\arguments{
\item{X}{x}

\item{Y}{x}

\item{lambdas}{x}

\item{nonzeros}{x}

\item{label}{x}

\item{cv_n_folds}{x}

\item{penalization}{x}

\item{max_iteration}{x}

\item{tolerance}{x}
}
\value{
x
}
\description{
Title
}
\examples{
x
}
