% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srda_core.R
\name{.srda_core}
\alias{.srda_core}
\title{The core algorithm of the sparse redundancy analysis (sRDA)}
\usage{
.srda_core(X, Y, lambda, nonzero, penalization, max_iteration, tolerance)
}
\arguments{
\item{X}{A explanatory matrix or data frame;}

\item{Y}{A response matrix or data frame;}

\item{lambda}{The ridge penalty parameter of the explanatory latent variable (Xi) used for enet}

\item{nonzero}{The number of non-zero alpha weights of the explanatory latent variable (Xi) used for enet or ust}

\item{penalization}{The penalization method: "enet", "ust", "none"}

\item{max_iteration}{Maximum number of iterations}

\item{tolerance}{Convergence criteria}
}
\value{
A list
}
\description{
This is an internal function.
}
\examples{
testing
}
