% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sRDA.R
\name{predict.sRDA}
\alias{predict.sRDA}
\title{S3 methods predict
predict correlations of latent dimensions}
\usage{
\method{predict}{sRDA}(rda, exploratory_new, response_new, ...)
}
\arguments{
\item{rda}{x}

\item{exploratory_new}{y}

\item{response_new}{z}

\item{...}{x}
}
\value{
x
}
\description{
S3 methods predict
predict correlations of latent dimensions
}
