% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_boostrapping_test.R
\name{run_bootstrapping_test}
\alias{run_bootstrapping_test}
\title{Bootstrapping test for the significance of features in sRDA}
\usage{
run_bootstrapping_test(explanatory, response, lambda, nonzero, n_bootstrap, ci)
}
\arguments{
\item{explanatory}{x}

\item{response}{x}

\item{lambda}{x}

\item{nonzero}{x}

\item{n_bootstrap}{Number of bootstrapping}

\item{ci}{Confidence interval}
}
\value{
a result data frame
}
\description{
Bootstrapping test for the significance of features in sRDA
}
\examples{
x
}
