% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_permutation_test.R
\name{run_permutation_test}
\alias{run_permutation_test}
\title{Permutation test for the significance of features in sRDA}
\usage{
run_permutation_test(rda, n_permutation, plot = FALSE)
}
\arguments{
\item{rda}{x}

\item{n_permutation}{x}

\item{plot}{x}
}
\value{
a result data frame
}
\description{
Permutation test for the significance of features in sRDA
}
\examples{
x

}
