% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srda.R
\name{srda}
\alias{srda}
\title{Sparse Redundancy Analysis (sRDA)}
\usage{
srda(
  explanatory,
  response,
  penalization,
  lambdas = 1,
  nonzeros = 1,
  max_iteration = 100,
  tolerance = 1e-20,
  parallel = FALSE,
  cv_n_folds = 5,
  n_lvs = 1,
  seed
)
}
\arguments{
\item{explanatory}{x}

\item{response}{x}

\item{penalization}{x}

\item{lambdas}{x}

\item{nonzeros}{x}

\item{max_iteration}{x}

\item{tolerance}{x}

\item{parallel}{x}

\item{cv_n_folds}{x}

\item{n_lvs}{x}

\item{seed}{x}
}
\value{
x
}
\description{
Sparse Redundancy Analysis (sRDA)
}
\examples{
x
}
