% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_paraellel_cv.R
\name{.run_parallel_cv}
\alias{.run_parallel_cv}
\title{run cross-validation with parallel processing}
\usage{
.run_parallel_cv(
  X,
  Y,
  lambdas,
  nonzeros,
  label,
  cv_n_folds,
  penalization,
  max_iteration,
  tolerance
)
}
\arguments{
\item{X}{explanatory matrix (n x p)}

\item{Y}{response matrix (n x q)}

\item{lambdas}{hyperparameter(s) for penalization}

\item{nonzeros}{maximum number of non-zero coefficients}

\item{label}{column labels of the X}

\item{cv_n_folds}{number of folds for cross-validation}

\item{penalization}{which penalization method}

\item{max_iteration}{maximum number of iterations}

\item{tolerance}{tolerance}
}
\value{
x
}
\description{
This is an internal function.
TODO Performance issues: the initiation is slow
}
\examples{
x
}
