% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srda.R
\name{srda}
\alias{srda}
\title{Sparse Redundancy Analysis (sRDA)}
\usage{
srda(
  explanatory,
  response,
  penalization,
  lambdas = 1,
  nonzeros = 1,
  n_lvs = 1,
  max_iteration = 100,
  tolerance = 1e-06,
  parallel = FALSE,
  cv_n_folds,
  seed
)
}
\arguments{
\item{explanatory}{A explanatory matrix or data frame}

\item{response}{A response matrix or data frame}

\item{penalization}{The penalization method: "enet", "ust", "none"}

\item{lambdas}{The ridge penalty parameter of the explanatory latent variable (Xi) used for enet}

\item{nonzeros}{The number of non-zero alpha weights of the explanatory latent variable (Xi) used for enet or ust}

\item{n_lvs}{x}

\item{max_iteration}{Maximum number of iterations}

\item{tolerance}{Convergence criteria}

\item{parallel}{Run the cross-validation in parallel (logical; TRUE or FALSE)}

\item{cv_n_folds}{x}

\item{seed}{x}
}
\value{
A sRDA object
}
\description{
This package is a rewrite of the sRDA (https://cran.r-project.org/package=sRDA) package available on CRAN
Please visit the original package and their paper for more details.
If you use this package, please consider cite the original paper of sRDA.
}
\examples{
x
}
